PlotlyFrontend/tammy_sample_app/requirements.txt
dash==2.3.1
pandas==1.3.0
python-dotenv==0.20.0
databricks-sql-connector==2.0.2
pyspark==3.2.1
numpy
dash-mantine-components==0.10.2
